requirements.txt
python==3.8.5
kaggle==1.5.12
numpy==1.22.3 
pandas==1.4.1 
matplotlib==3.5.1  
scikit-learn==1.0.2 
scipy==1.8.0
seaborn==0.11.2 
statsmodels==0.13.2 
xgboost==1.5.1
pyyaml==6.0
pytest==7.0.1